ParseException.java: move to common directory for parser
It does not make sense to put the this exception in a separate directory.
