+++ src/main/java/vimification/logic/parser/DeleteCommandParser.java
@@ -4,7 +4,6 @@ import static vimification.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import vimification.commons.core.index.Index;
 import vimification.logic.commands.DeleteCommand;
-import vimification.logic.parser.exceptions.ParseException;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
@@ -14,7 +13,7 @@ public class DeleteCommandParser implements Parser<DeleteCommand> {
     /**
      * Parses the given {@code String} of arguments in the context of the DeleteCommand and returns
      * a DeleteCommand object for execution.
-     * 
+     *
      * @throws ParseException if the user input does not conform the expected format
      */
     public DeleteCommand parse(String args) throws ParseException {
