+++ src/main/java/vimification/logic/parser/EditCommandParser.java
@@ -16,7 +16,6 @@ import java.util.Set;
 import vimification.commons.core.index.Index;
 import vimification.logic.commands.EditCommand;
 import vimification.logic.commands.EditCommand.EditPersonDescriptor;
-import vimification.logic.parser.exceptions.ParseException;
 import vimification.model.tag.Tag;
 
 /**
@@ -27,7 +26,7 @@ public class EditCommandParser implements Parser<EditCommand> {
     /**
      * Parses the given {@code String} of arguments in the context of the EditCommand and returns an
      * EditCommand object for execution.
-     * 
+     *
      * @throws ParseException if the user input does not conform the expected format
      */
     public EditCommand parse(String args) throws ParseException {
