+++ src/main/java/vimification/logic/parser/FindCommandParser.java
@@ -5,7 +5,6 @@ import static vimification.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import java.util.Arrays;
 
 import vimification.logic.commands.FindCommand;
-import vimification.logic.parser.exceptions.ParseException;
 import vimification.model.person.NameContainsKeywordsPredicate;
 
 /**
@@ -16,7 +15,7 @@ public class FindCommandParser implements Parser<FindCommand> {
     /**
      * Parses the given {@code String} of arguments in the context of the FindCommand and returns a
      * FindCommand object for execution.
-     * 
+     *
      * @throws ParseException if the user input does not conform the expected format
      */
     public FindCommand parse(String args) throws ParseException {
