+++ src/main/java/vimification/logic/parser/Parser.java
@@ -1,7 +1,6 @@
 package vimification.logic.parser;
 
 import vimification.logic.commands.Command;
-import vimification.logic.parser.exceptions.ParseException;
 
 /**
  * Represents a Parser that is able to parse user input into a {@code Command} of type {@code T}.
@@ -10,7 +9,7 @@ public interface Parser<T extends Command> {
 
     /**
      * Parses {@code userInput} into a command and returns it.
-     * 
+     *
      * @throws ParseException if {@code userInput} does not conform the expected format
      */
     T parse(String userInput) throws ParseException;
